revert merge diff
What is needed is structural conflicts handling, i.e. if { someProp: "string" } is in conflict with { someProp: { sub: "property" } } then someProp.sub should also be considered conflicting. This logic doesn't belong to Diff/Merge level but must be handled at a level that is aware of the actual structure of the data (e.g. Java/Json). Another possibility might be versioning SortedMaps instead of Maps with keys that in addition to compare would have startsWith/isChildOf method to determine not only the order but also structure of the keys. Merge/Diff would benefit from that kind of keys and SortedMap. Also list binding would be more efficient and simpler if one could rely on index order... But then again that kind of requirements on keys would have it's own penalty and would only work with strictly hierarchical data and not with object graphs/references.